requirements.txt
Django>=5.2
djangorestframework